add main file with options and make user inpute a choose
